requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
appdirs==1.4.4
asgiref==3.5.2
astor==0.8.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
Automat==20.2.0
Babel==2.12.1
bce-python-sdk==0.8.79
bcrypt==3.2.0
beautifulsoup4==4.11.1
Bottleneck==1.3.5
branca==0.6.0
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
constantly==15.1.0
cryptography==38.0.4
cssselect==1.1.0
cycler==0.11.0
datasets==2.10.1
decorator==5.1.1
dill==0.3.4
Django==3.2.15
easydict==1.10
echarts-china-cities-pypkg==0.0.9
echarts-china-counties-pypkg==0.0.2
echarts-china-misc-pypkg==0.0.1
echarts-china-provinces-pypkg==0.0.3
echarts-countries-pypkg==0.1.6
entrypoints==0.4
fastapi==0.95.0
ffmpy==0.3.0
filelock==3.9.0
fire==0.5.0
Flask==2.2.3
Flask-Babel==2.0.0
flit_core==3.6.0
folium==0.14.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gradio==3.22.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.2
hyperlink==21.0.0
icon-font-to-png==0.4.1
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
incremental==21.3.0
itemadapter==0.3.0
itemloaders==1.0.4
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==2.0.0
lml==0.0.2
lxml==4.9.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.4.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.4
multiprocess==0.70.12.2
numexpr==2.8.4
numpy==1.21.5
opencv-python==4.7.0.72
opt-einsum==3.3.0
orjson==3.8.7
packaging==22.0
paddle-bfloat==0.1.7
paddle2onnx==1.0.6
paddlefsl==1.1.0
paddlehub==2.3.1
paddlenlp==2.5.2
paddlepaddle==2.4.2
palettable==3.3.0
pandas==1.3.5
parsel==1.6.0
Pillow==9.4.0
pip==22.3.1
pkgutil_resolve_name==1.3.10
prettytable==3.6.0
Protego==0.1.16
protobuf==3.19.6
psycopg2==2.9.3
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.6
PyDispatcher==2.0.5
pydub==0.25.1
pyecharts==2.0.2
pyecharts-jupyter-installer==0.0.3
pyecharts-snapshot==0.2.0
pyee==8.2.2
Pygments==2.14.0
PyHamcrest==2.0.2
PyMySQL==1.0.2
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2022.7
pywin32==305.1
PyYAML==6.0
pyzmq==25.0.1
queuelib==1.5.0
rarfile==4.0
requests==2.28.1
requests-file==1.5.1
responses==0.18.0
rfc3986==1.5.0
rich==13.3.2
scikit-learn==1.0.2
scipy==1.7.3
Scrapy==2.6.2
sentencepiece==0.1.97
seqeval==1.2.2
service-identity==18.1.0
setuptools==65.6.3
simplejson==3.18.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
starlette==0.26.1
stylecloud==0.5.2
termcolor==2.2.0
threadpoolctl==3.1.0
tinycss==0.4
tldextract==3.2.0
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
Twisted==22.2.0
twisted-iocpsupport==1.0.2
typer==0.7.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
urllib3==1.26.14
uvicorn==0.21.1
visualdl==2.4.2
w3lib==1.21.0
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.2.3
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
wordcloud==1.8.2.2
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
zope.interface==5.4.0
